BJS2-97290 я победил сей баг!
