+++ src/main/java/school/faang/user_service/dto/kafka/EventStartEventDto.java
@@ -1,5 +1,6 @@
 package school.faang.user_service.dto.kafka;
 
+import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.service.event.scheduling.TimeLeft;
 
 import java.util.List;
@@ -8,7 +9,7 @@ public record EventStartEventDto(
         Long eventId,
         Long userId,
         String nameOwner,
-        List<Long> attendeesIds,
+        List<UserDto> attendeesUser,
         String titleEvent,
         TimeLeft timeLeft
 ) {
