+++ src/main/java/school/faang/user_service/repository/user/UserRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import school.faang.user_service.entity.user.User;
 import school.faang.user_service.exception.EntityNotFoundException;
 
@@ -37,7 +38,15 @@ public interface UserRepository extends JpaRepository<User, Long> {
                 .orElseThrow(() -> new EntityNotFoundException(String.format("User %d not found", userId)));
     }
 
+    List<User> findAllByIdIn(List<Long> userIds);
 
     Page<User> findByIdNotIn(List<Long> userIds, Pageable pageable);
+
+    @Query(value = """
+            SELECT u.* FROM users u 
+            JOIN user_event ue ON u.id = ue.user_id 
+            WHERE ue.event_id = :eventId
+            """, nativeQuery = true)
+    List<User> findAttendeesByEventId(@Param("eventId") long eventId);
 
 }
